smart-camera/app/requirements.sound.txt
pygame==2.1.*
paho-mqtt==1.5.*
